Move dashboard UI select option types to input types and re-use for combo select option #3244.
